Use time.Timer instead of time.After
memory is allocated every time time.After is called and is not freed
until the timer expires. This could create a memory "leak" of sorts if
the user of cron calls snapshots or adds new entries frequently.

using time.Timer allows for explicitly stopping and cleaning up the
timer on each iteration of the for-loop. So when a snapshot or new entry
is added, we can cleanup the timer.

see https://groups.google.com/forum/#!topic/golang-nuts/cCdm0Ixwi9A for
more details.